Refactor add-contacts & add-pending-contact
We wanted to move towards having multiple functions
manipulating the data rather than relying on a chain of
dispatched events.

I have refactored ui.screens.contacts.events so that now the
functions mainly manipulate fx, passed as a first parameter.

Their responsability is to make sure their fx are merged in, so
that they can be composed more easily.

Signed-off-by: Eric Dvorsak <eric@dvorsak.fr>