[#4755] Keep the transaction source intact on errors/discards.
If a tx was send from a chat, that fact is recorded in its properties, `from-chat?` property.
On errors and discards, this property was reset to `false`, hence, making any chat transaction editable if an error happens or user discards it.
This commit makes this property stay as it is.

Signed-off-by: Igor Mandrigin <i@mandrigin.ru>