Update bundle script
With the change in status-go mobile the name of the artifact as changed
and currently the script is looking at the wrong file.
I have also added an option to not clean the repo (only if specified),
as I rarely want to clean the repo myself, and I have always to comment
it out.

Signed-off-by: Andrea Maria Piana <andrea.maria.piana@gmail.com>