Copy update multiaccount
Fixes copy update in #8447

In the UI we want to refer to a watch only address as `address` instead of `account`. Interactions with an address are limited compared to an account in wallet. e.g. there is no option to send assets from this account as Status wallet does not have the key to sign for transactions.

Signed-off-by: Andrey Shovkoplyas <motor4ik@gmail.com>