fix(multiaccounts/login): ensure log-level is updated correctly
As reported in #10127 there's an issue that the log level isn't properly
applied when changed through log settings UI.

Turned out that we're setting the log-level only once in the entire
lifetime of the application and that is during bootstrap.
When the log level is changed through the UI, we never actually update
the log level in the logger module.

To fix this, we just need to ensure that, whenever the log-level is changed,
we also call `log/set-level!` with the new level.

It should also be noted that, every time the log level is changed through
the UI, the user is logged out, so one place to update the logger is
after very successful login.

Fixes #10127

Signed-off-by: Andrey Shovkoplyas <motor4ik@gmail.com>